+++ src/ui/components/App.tsx
@@ -1,5 +1,6 @@
 import React from "react"
 import { Provider } from "react-redux"
+import { ToastContainer } from "react-toastify"
 import { PersistGate } from "redux-persist/integration/react"
 import { configureStore } from "../store/configureStore"
 import { StackRouter } from "./utils/StackRouter"
@@ -10,6 +11,18 @@ export const App = () => {
     <Provider store={store}>
       <PersistGate loading={null} persistor={persistor}>
         <StackRouter />
+        <ToastContainer
+          position="top-right"
+          autoClose={3000}
+          hideProgressBar={false}
+          newestOnTop={false}
+          draggablePercent={60}
+          closeOnClick
+          rtl={false}
+          pauseOnVisibilityChange
+          draggable
+          pauseOnHover
+        />
       </PersistGate>
     </Provider>
   )